# internal/handlers/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Relational Knowledge Engineering Platform API
  description: |
    A comprehensive API for document processing, knowledge extraction, and graph-based analysis.
    
    The platform supports:
    - PDF document upload and processing
    - Chunked file uploads for large documents
    - Text extraction and NLP processing
    - Knowledge graph generation and analysis
    - Document management and retrieval
    
    ## Authentication
    Currently, the API does not require authentication for public endpoints.
    
    ## Rate Limiting
    No rate limiting is currently implemented.
    
    ## Error Handling
    All endpoints return consistent error responses with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Relational Knowledge Engineering Platform
    url: https://github.com/yothgewalt/relational-knowledge-engineering-platform-server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/documents:
    get:
      tags: [Documents]
      summary: List documents
      description: |
        Retrieves a list of all uploaded documents with their metadata and processing status.
      operationId: listDocuments
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of documents to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of documents to skip
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProcessingStatus'
          description: Filter by processing status
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/documents/{id}:
    get:
      tags: [Documents]
      summary: Get document by ID
      description: |
        Retrieves a specific document by its ID, including all extracted text,
        processing results, and metadata.
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags: [Documents]
      summary: Delete document
      description: |
        Deletes a document and all associated data including extracted text,
        graph data, and the original file. This operation cannot be undone.
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Document deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/graphs/{id}:
    get:
      tags: [Graphs]
      summary: Get knowledge graph
      description: |
        Retrieves the knowledge graph network for a specific document,
        including nodes, edges, and their relationships.
      operationId: getGraphNetwork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Graph network retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphNetwork'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          example: "doc_123e4567-e89b-12d3-a456-426614174000"
        filename:
          type: string
          example: "research-paper.pdf"
        content_type:
          type: string
          example: "application/pdf"
        size:
          type: integer
          format: int64
          example: 52428800
        uploaded_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
        processed_at:
          type: string
          format: date-time
          nullable: true
          example: "2023-12-01T10:05:00Z"
        status:
          $ref: '#/components/schemas/ProcessingStatus'
        extracted_text:
          $ref: '#/components/schemas/ExtractedText'
        processed_text:
          $ref: '#/components/schemas/ProcessedText'
        graph_id:
          type: string
          nullable: true
          example: "graph_123e4567-e89b-12d3-a456-426614174000"
        error_message:
          type: string
          nullable: true
          example: null
    ProcessingStatus:
      type: string
      enum: [uploaded, processing, completed, error]
      description: Current processing status of the document
    ExtractedText:
      type: object
      properties:
        content:
          type: string
          description: Raw extracted text from the document
        metadata:
          type: object
          description: Additional extraction metadata
        page_count:
          type: integer
          description: Number of pages processed
    ProcessedText:
      type: object
      properties:
        sentences:
          type: array
          items:
            $ref: '#/components/schemas/Sentence'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
          description: Generated summary of the document
    Sentence:
      type: object
      properties:
        text:
          type: string
          description: The sentence text
        start_offset:
          type: integer
          description: Character offset where sentence starts
        end_offset:
          type: integer
          description: Character offset where sentence ends
        sentiment:
          type: number
          format: float
          minimum: -1
          maximum: 1
          description: Sentiment score (-1 to 1)
    Entity:
      type: object
      properties:
        text:
          type: string
          description: The entity text
        label:
          type: string
          description: Entity type/label
        start_offset:
          type: integer
          description: Character offset where entity starts
        end_offset:
          type: integer
          description: Character offset where entity ends
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score (0 to 1)
...
    GraphNetwork:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        metadata:
          type: object
          properties:
            total_nodes:
              type: integer
            total_edges:
              type: integer
            creation_time:
              type: string
              format: date-time
    GraphNode:
      type: object
      properties:
        id:
          type: string
          description: Unique node identifier
        label:
          type: string
          description: Node label/name
        type:
          type: string
          description: Node type (entity, concept, etc.)
        properties:
          type: object
          additionalProperties: true
          description: Additional node properties
        weight:
          type: number
          format: float
          description: Node importance weight
    GraphEdge:
      type: object
      properties:
        id:
          type: string
          description: Unique edge identifier
        source:
          type: string
          description: Source node ID
        target:
          type: string
          description: Target node ID
        relationship:
          type: string
          description: Type of relationship
        weight:
          type: number
          format: float
          description: Edge weight/strength
        properties:
          type: object
          additionalProperties: true
          description: Additional edge properties
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
          nullable: true
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: 404
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: 500
...